pyproject.toml
[tool.poetry]
name = "detectiq"
version = "0.1.0"
description = "A detection engineering workbench with LLM capabilities"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "detectiq"}]


[tool.poetry.dependencies]
python = ">=3.9, <3.14"
sigmaiq = "^0.4.5"
langchain = "^0.3.7"
langchain-community = "^0.3.7"
langchain-openai = "^0.2.8"
langchain-core = "^0.3.17"
openai = "^1.54.4"
tiktoken = "^0.8.0"
faiss-cpu = "^1.9.0"
yara-python = "^4.5.1"
dpkt = "^1.9.8"
scapy = "^2.6.1"
cryptography = "^42.0.5"
python-dotenv = "^1.0.1"
plyara = "^2.1.1"
requests = "^2.32.3"
aiofiles = "^24.1.0"
idstools = "^0.6.5"
djangorestframework-simplejwt = "^5.3.1"
pefile = "^2024.8.26"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-cov = "^5.0.0"
black = "^24.8.0"
ruff = "^0.7.2"
pytest-asyncio = "^0.24.0"
python-dotenv = "^1.0.1"

[tool.poetry.group.splunk]
optional = true

[tool.poetry.group.splunk.dependencies]
splunk-sdk = "^2.1.0"

[tool.poetry.group.elastic]
optional = true

[tool.poetry.group.elastic.dependencies]
elasticsearch = {extras = ["async"], version = "^8.0.0"}

[tool.poetry.group.microsoft]
optional = true

[tool.poetry.group.microsoft.dependencies]
msal = "^1.31.1"

[tool.poetry.group.webapp]
optional = true

[tool.poetry.group.webapp.dependencies]
django = "^4.2.0"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.0.0"
psycopg2-binary = "^2.9.0"
django-environ = "^0.10.0"
pyyaml = "^6.0"
djangorestframework-simplejwt = "^5.3.1"
django-extensions = "^3.2.3"

[tool.poetry.group.dev]
optional = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py311']

[tool.ruff]
line-length = 120  # Same as Black
# Assume Python 3.11
target-version = "py311"
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
    ".venv"
]

[tool.ruff.lint]
# Enable isort rules
select = ["I"]

# Sort imports
[tool.ruff.lint.isort]
known-first-party = ["detectiq"]
combine-as-imports = true
case-sensitive = false


[tool.coverage.run]
omit = [
    # omit anything in a tests directory
    "tests/*",
    ]

# Pytest configuration
[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "--cov=detectiq --cov-report=term-missing"
asyncio_mode = "auto"

[tool.poetry.scripts]
update-reqs = "update_requirements:main"